Decision - Oath of Eorl WIP
Implements the Oath of Eorl decision for Gondor and Rohan, allowing them to call for aid from the other when they are outnumbered in war.  Possible aid includes direct intervention, event spawned troops and refusal.

Still need to add fringe cases such as disabling if at war with the other party and adding in AI weights.