Continue WIP Orc names
continuation the name list for the orc bois, there is now 3 list of orc names, first one is the current one that existed, which i think we'll be the more mordor orc, secound is just generic orcs in the world. and the 3d one is mountain orcs. The localisation is not completed yet