Introduce heir decision events for Variags
For #498

This commit should bring the Introduce Heir decision for Variags to a testable state.  We currently have the bare minimum number of trials in there, but as it's a pool, more can easily be added, much like the sailing events.  And the framework should be clear enough to make it obvious how to structure these.

Ready to test!