WIP #705, finished 1/3 decisions so far
The decision for the ruler of Anfalas to invite elves to Edhellond is finished, just needs testing when compatch is stable enough to allow for a few years to be run.
To debug the first decision, i.e., invite elves to Edhellond, MAKE SURE you are friends with one of the elven rulers as that is a prerequisite for the decision, otherwise the event won't fire properly.

The decision for Aragorn/Gondor to elevate Edhellond into a kingdom tiered title still needs flavour text and needs a bit more fleshing out in terms of what it actually does.

The decision for independent elves to elevate Edhellond into a kingdom is barebones, basically nothing has been done there yet,