"Temporarily" disables Dreams of Distant Lands  wanderer lifestyle event
"Temporarily" disables the dreams of distant lands wanderer lifestyle event. (wanderer_lifestyle_events.30)

Someone previously tried to compatch it, but only did so partially and did not consider the actual trigger setup and customizable localization attached to it, which heavily references vanilla geographical regions.

In order to fix this, we have to define proper equivalencies for vanilla regions (which is long overdue anyway)

#1211