Adds Khäz-Gramaze Missionary LAAMP to LW
Adds a Khäz-Gramaze Missionary LAAMP to LW
Currently wont convert people, just a scholar camp.

Cousin of the LW King, orphaned in his youth and sent to the Temple.
Historically attempts to spread the cult of Khäz-Gramaze among the Jelut and Tedjin.
Is unsuccessful in gaining many supporters and slain by first ruler of Tedjin confederacy