New Arnor Faf
Did this stuff before it was put on hold again, but I honestly don't want it to just sit on my PC or to get rid of it, as both of those will bum me out for days. So, here it is, commented out.
If we decide against, I will gladly remove it, but I can't stand just sitting on it.
It is very revertible too, so if you can't stand having it exist on the Dev Version at all, even commented out, just revert it. At least I don't have to do it and cause myself grief.