Adds Cochineal Industry Building and Decision
Implements the cochineal industry from RICE!

Adds a Cochineal Industry regional duchy building, as well as a minor decision to further develop it (with a stronger buff in duchies that have the building).

These little bugs are capable of producing marvellous red dye, to be exported or used domestically.

Currently two species are in the mod: One is native in Logathavuld and around the eastern Sea of Rhûn, another lives exclusively around the Sea of Núrn.