House of the First Elder Done
Changed the colours of the Original and Avari Elves, so they did not have any overlaps.
Renamed the Tatyarin Ethnicity to actually be tatyarin_ethnicity.
Fixed Sekak culture colour trigger, because I was tired of that one letter making me see Magenta.
Also slight adjustment to the evil alignment, because it was looking really odd with the new lighting.